test: Add verification of the nautilus cts api.
Verify we can call GetCapabilities, and that it returns a
textgroup with our test collection urn.

Verify we can call GetPassage with the first work urn from
the GetCapabilities query and that it has the proper prefix
to its own urn.

Both of these will fail with an empty corpus, which maybe
shouldn't be an error. Could mock the app with an empty
corpus and check that, once the nautilus resolver issue
with empty corpora is fixed?