Add a link to /sciafe/ if it's available.
On the test server, we have a viewer webapp available on this
path, and it's helpful to link to it from the landing page since
it's the most accessible interface.

However, it's as separate repos and not available by default.
To deal with this, ping it at template render time and add
the link only if it's resolving. This is kind of a terrible
hack in that it doubles traffic, but will do for now.