Set scaife link url from a config file.
Replace the hack of pinging a hard-coded url for an installed
scaife viewer with loading a url from an optional config file.

This is more flexible, and efficient, but the query also failed
in the test environment.

On a deployment server which does provide the viewer, add a file
named `config.py` with a definition of the correct path.

For example:

    SCAIFE_URL = 'scaife/'

If the file isn't present or doesn't define the key, the link
will not be present.